+++ Controls-demo/list_new/Navigation/Cut/CutNavigation.less
@@ -1,7 +1,3 @@
 .controls-demo__List-Cut_containerWidth {
     width: 300px;
 }
- 
-.controls-demo__List-Cut_columnHeader {
-    font-weight: bolder;
-}
