+++ Controls-demo/list_new/Navigation/Cut/ListCut/ListCut.wml
@@ -1,0 +1,3 @@
+<Controls.list:View
+    source="{{_source}}"
+    navigation="{{_navigation}}"/>
